Add landing page (!3)
This patch adds a landing page at the root path. A customer arriving at this
page may click on the start button to load the preregistration form. If all the
statements on this form are not answered affirmatively (by ticking the boxes)
then the customer is shown the form again with corresponding errors. Otherwise
they are redirected to a blank page.

This patch is a union of the following commits:

1. Updates Ruby to 2.3.1

Removes a bunch of unneeded gems and also appeases Rubocop

2. Add previously removed gems

The following gems were added back to the project as the suspenders version of
Rails is dependent upon them:

  + flutie
  + high_voltage
  + title

3. Remove style gems

These will not be required when we use the Gov.UK style gems.

4. Clean up Gemfile
5. Add Gov.UK style gems
6. Use Gov.UK stylesheets
7. Remove normalize-rails style gem
8. Style layout to look like Gov.UK website
9. Add static start page
10. Add preregistration model
11. Add preregistrations controller
12. Refactor preregistration specs
13. Add view for preregistration form
14. Correctly style form for Gov.UK website
15. Connect start button to preregistration form
16. Still not using canonical host...
17. Correct name of MCA
18. Add copyright information to footer
19. Add hack for start button icon
20. Add alpha phase banner to page layout
21. Amend preregistration form submit button text
22. Add feature spec