Prefer #touch over callbacks for payment/declarations
- We need a way to move a submission from :incomplete
to :unassigned when a payment or declaration is made. 

- Now that we can accept payments and declarations
for submissions that aren’t incomplete, we need a
clean way for submissions that are already assigned
to skip the transition to :unassigned (because
it will break!)

- As payments and declarations:
belong_to :submission, touch: true
there is and after_touch callback on the 
submission that checks whether the state can be
changed.