Create notifications even if they can’t be sent
- This might be a questionable strategy

If there is no email address, we are still creating
the notification object. Within the #send method
there is a guard that checks whether the notification
is deliverable?

The idea is that, while the notification won’t be
sent, there will be an audit trail that someone
tried to send the email. We can then inspect the notification
record and see whether it could have been delivered.