Build a Closed Registration
- Here we insert a new record with the closed_at
date and message

- And we insert the previous registration dates
to the new record. Note the `try` wrapper. Taking
a defensive approach as I know that, somewhere
down the line we are going to be trying to
close vessels that don’t have a current-registration.
This should prevent some poor person cursing my
over-confident code!

- This makes sense because a Closure is the `state`
of a current registration