Historic transcripts don’t display duplicates
- When we built the historic transcripts,
we were asking for all the registration instances.
But this produces lots of duplicates as we record a "registration"
every time a submission is approved by a reg officer

- What we really want is all a list of all unique registration dates
and the most recent changeset for those registrations.

- The implementation (vessel#historic_registrations) is a little clunky,
but we won't be dealing with more than a handful of records at a time
so it should be sufficient