Bug fix: electronic_delivery PDFs were blank
- Initially we said that submissions with
electronic_delivery would not go through the 
application processor

- However, this meant that when the 
transcript was generated, there was not registration
object with which to build it. So it was empty

- Here we ensure that electronic_delivery
applications still go through the application
processor but we make sure that print_jobs
are not built. There might come a time when we
want this functionality

- A word too about creating a record for every 
application in the registration table. It might
sound daft - why would be create registrations for
transcript requests? Well, the requirements were
pretty vague as to what happened when a submission
was processed. RSS wanted to be able to change this
or that, so we needed a way to capture whatever
changes were made to the registry