Introduce a state :initialising
- An excerpt from the README:

When a submission enters the default :incomplete state, 
it fires an event `build_default`, invoking the 
`SubmissionBuilder` and setting up all the defaults. 

- If you want to initialize a submission object to 
build a form or any other instance when you don't 
want a record to be created, you need to forecfully 
set the state to :initializing. 
For example: `Submission.new(state: :initializing)`.